% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation_nivsupra.R
\name{agregation_nivsupra}
\alias{agregation_nivsupra}
\title{Titre}
\usage{
agregation_nivsupra(table_entree, code_com = colnames(table_entree)[1],
  var_num = colnames(table_entree[, sapply(table_entree, is.numeric)]),
  nivsupra, na.nivgeo.rm = F)
}
\value{
Remplir
}
\description{
Remplir
}
