% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changement_COG_varCar.R
\name{changement_COG_varCar}
\alias{changement_COG_varCar}
\title{Titre}
\usage{
changement_COG_varCar(table_entree, annees,
  codgeo_entree = colnames(table_entree)[1],
  methode_fusion = "methode_difference", mot_different = "différents")
}
\value{
Remplir
}
\description{
Remplir
}
