% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_table_passage_inverse.R
\name{creer_table_passage_inverse}
\alias{creer_table_passage_inverse}
\title{Titre}
\usage{
creer_table_passage_inverse(table_passage, annee_debut = substr(table_passage,
  9, 12), donnees_insee = ifelse(length(grep("insee", psg)) == 1, T, F))
}
\value{
Remplir
}
\description{
Remplir
}
