% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changement_COG_typo.R
\name{changement_COG_typo}
\alias{changement_COG_typo}
\title{Titre}
\usage{
changement_COG_typo(table_entree, annees,
  codgeo_entree = colnames(table_entree)[1],
  var_car = colnames(table_entree)[-which(colnames(table_entree) ==
  codgeo_entree)], methode_fusion = "methode_difference",
  mot_different = "différents", donnees_insee = T, libgeo = NULL)
}
\value{
Remplir
}
\description{
Remplir
}
