% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changement_COG_typo_details.R
\name{changement_COG_typo_details}
\alias{changement_COG_typo_details}
\title{Titre}
\usage{
changement_COG_typo_details(table_entree, annees,
  codgeo_entree = colnames(table_entree)[1], typo,
  methode_fusion = "methode_difference", mot_difference = NULL,
  classe_absorbante = NULL, donnees_insee = T)
}
\value{
Remplir
}
\description{
Remplir
}
