% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changement_COG_varNum.R
\name{changement_COG_varNum}
\alias{changement_COG_varNum}
\title{Titre}
\usage{
changement_COG_varNum(table_entree, annees,
  codgeo_entree = colnames(table_entree)[1],
  var_num = colnames(table_entree[, sapply(table_entree, is.numeric)]),
  clef_commune_unique = T, libgeo = NULL, donnees_insee = T)
}
\value{
Remplir
}
\description{
Remplir
}
