% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changement_COG_varNum.R
\name{changement_COG_varNum}
\alias{changement_COG_varNum}
\title{Changer les variables numériques de géographie communale}
\usage{
changement_COG_varNum(table_entree, annees,
  codgeo_entree = colnames(table_entree)[1],
  var_num = colnames(table_entree)[sapply(table_entree, is.numeric)],
  agregation = T, libgeo = NULL, donnees_insee = T)
}
\arguments{
\item{table_entree}{correspond à la table à transformer en une autre géographie}

\item{annees}{est un vecteur qui liste l'ensemble des années qui séparent le code officiel géographique de départ et d'arrivée. Par exemple c(1968:1985). Le package rend possible l'utilisation de tables de passages d'une année de COG vers une année antiérieure (par exemple c(2016:2014)).}

\item{codgeo_entree}{est une chaîne de caractères qui indique le nom de la variable contenant les codes Insee communaux. Par défaut, il s'agit du nom de la première colonne de table_entree.}

\item{var_num}{est un vecteur de chaînes de caractères qui indique les noms des variables numériques à convertir. Par défaut, il s'agit de l'ensemble des variables de types numériques dans table_entree.}

\item{agregation}{vaut TRUE si la table souhaitée doit sommer toutes les lignes qui concernent une même commune et FALSE si l'on souhaite volontairement conserver les doublons dans les codes commune (dans les tables de flux par exemple). Si agregation = F, les variables de type caractère sont alors conservées comme telles ou dupliquées en cas de défusion et les variables numériques sommées en cas de fusion ou réparties proportionnellement à la population de chaque commune en cas de défusion.}

\item{libgeo}{doit être conservé à NULL si la table souhaitée ne contient pas de libellés de communes, sinon il doit indiquer le nom de la variable renseignant sur le libellé de la commune (celui présent dans la table si il existe déjà ou tout autre nom s'il doit être ajouté).}

\item{donnees_insee}{vaut TRUE si les données manipulées sont produites par l'Insee. En effet, quelques rares modifications communales (la défusion des communes Loisey et Culey au 1er janvier 2014 par exemple) ont été prises en compte dans les bases de données communales de l'Insee plus tard que la date officielle. Pour tous les COG officiels datant d'avant 2008, seules les tables de passage Insee sont disponibles dans ce package.}
}
\description{
Transformer des tables de données numériques en géographie au premier janvier d'une année souhaitée.
}
\details{
Le code officiel géographique de référence du package est actuellement celui au 01/01/2017. Les données communales devront être dans ce COG pour être agrégées en niveaux supracommunaux (fonction nivsupra). \cr

Les autres codes officiels géographiques utilisés dans COGugaison sont les suivants :
 - COG 1968 : à partir du 01/03/1968 (donnees_insee=T obligatoirement)
 - COG 1975 : à partir du 20/02/1975 (donnees_insee=T obligatoirement)
 - COG 1982 : à partir du 04/03/1982 (donnees_insee=T obligatoirement)
 - COG 1990 : à partir du 05/03/1990 (donnees_insee=T obligatoirement)
 - COG 1999 : à partir du 08/03/1999 (donnees_insee=T obligatoirement)
 - COG 2008 à 2017 : à partir du 01/01 de chaque année. (donnees_insee=T ou F) \cr

Les différences entre les tables de passage Insee et non Insee sont les suivantes :
1982-03-03 (pris en compte par l'Insee seulement après le 04/03/1982): Flaignes-Havys (08169) est un rassemblement de Flaignes-Havys (08169), Havys (08221) [fusion simple].
2014-01-01 (pris en compte par l'Insee seulement au 01/01/2015) : Loisey (55298) s'est séparée en Loisey (55298), Culey (55138) [rétablissement].
1990-02-01 (pris en compte par l'Insee seulement après le 05/03/1990) : Le code commune de Oudon passe de 14624 à 14697 [changement de code dû à un changement de chef-lieu]. \cr
}
\examples{
## Exemple 1
# Ici, nous allons transformer les variables numériques de la table *exemple_pop* afin de récupérer les données de population et de superficie des communes au 1er janvier 2017 (au lieu de 2014).
exemple_popcom_COG2017_num <- changement_COG_varNum(table_entree=exemple_popcom,annees=c(2014:2017),agregation=T,libgeo="LIBGEO",donnees_insee=T)
head(exemple_popcom_COG2017_num)
## Exemple 2
# La fonction peut également s'appliquer à des tables de flux (codes communes pouvant comporter des doublons) grâce à l'option agregation = FALSE.
exemple_flux_COG2017 <- changement_COG_varNum(table_entree=exemple_flux,annees=c(2014:2017),codgeo_entree="COMMUNE",agregation=F,libgeo=NULL,donnees_insee=T)
exemple_flux_COG2017 <- changement_COG_varNum(table_entree=exemple_flux_COG2017,annees=c(2014:2017),codgeo_entree="DCLT",agregation=F,libgeo=NULL,donnees_insee=T)
}
\references{
[l'historique des géographies communales](https://www.insee.fr/fr/information/2666684#titre-bloc-11)
[les tables d'appartenance des communes aux différents niveaux géographiques](https://www.insee.fr/fr/information/2028028)
}
\seealso{
\link{changemnt_COG_typo}, \link{changement_COG_typo_details},  \link{COG_akinator}, \link{enlever_PLM}, \link{modification_Corse}, \link{modifications_communales},\link{nivsupra}
}
