% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modification_Corse.R
\name{modification_Corse}
\alias{modification_Corse}
\title{Titre}
\usage{
modification_Corse(table_entree, codgeo_entree = colnames(table_entree)[1],
  vecteur_entree = is.vector(table_entree), sens = "20vers2A2B")
}
\value{
Remplir
}
\description{
Remplir
}
