% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modification_Corse.R,
%   R/modifications_communales.R
\name{modifications_communales}
\alias{modifications_communales}
\alias{modification_Corse}
\alias{modifications_communales}
\title{Titre}
\usage{
modification_Corse(table_entree, codgeo_entree = colnames(table_entree)[1],
  vecteur_entree = is.vector(table_entree), sens = "20vers2A2B")

modifications_communales(date_debut, date_fin)
}
\value{
Remplir

Remplir
}
\description{
Remplir

Remplir
}
