% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nivsupra.R
\name{nivsupra}
\alias{nivsupra}
\title{Titre}
\usage{
nivsupra(table_entree, code_com = colnames(table_entree)[1],
  var_num = colnames(table_entree)[sapply(table_entree, is.numeric)],
  nivsupra, nivsupra_nom = ifelse(agregation == T, nivsupra, paste0(nivsupra,
  "_", code_com)), na.nivgeo.rm = F, agregation = T)
}
\value{
Remplir
}
\description{
Remplir
}
